ci: Scan TypeScript with CodeQL (#27)
# Pull Request

## Description

This pull request includes updates to the CodeQL analysis workflow
configuration in the `.github/workflows/code-checks.yml` file. The
changes improve the flexibility and accuracy of the CodeQL analysis by
introducing a strategy matrix and updating the action versions.

Key changes include:

### Workflow Configuration Updates:

* Added `statuses: write` permission to the CodeQL analysis job.
* Introduced a strategy matrix to run the analysis for multiple
languages (`typescript` and `actions`).
* Updated the checkout action version to `v4.2.2` and capitalized the
step name for consistency.
* Updated the CodeQL action versions to `v3.28.9` for both
initialization and analysis steps.
* Modified the `languages` parameter in the CodeQL initialization step
to use the matrix language.

Fixes #9