ci: Check Justfile Format (#29)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for checking the format of the `Justfile`. The most important change is
the addition of the `check-justfile-format` job in the
`.github/workflows/code-checks.yml` file.

New job added to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R74-R87):
Added `check-justfile-format` job to ensure the `Justfile` format is
correct. This job checks out the repository, sets up Just, and runs the
format check.

Fixes #21