chore: Remove CodeQL Configuration (#56)
# Pull Request

## Description

This pull request updates the CodeQL configuration and actions used in
the repository to improve security and maintain compatibility with the
latest versions. The most important changes include removing an unused
query filter from the CodeQL configuration file and updating the CodeQL
action versions in the workflow file.

### CodeQL Configuration Updates:
*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Removed the `actions/unpinned-tag` query filter, as it is no longer
needed.

### Workflow Updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L110-R110):
Updated the `upload-sarif`, `init`, and `analyze` steps to use
`github/codeql-action` version `v3.28.15` instead of older versions.
This ensures compatibility with the latest CodeQL features.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L110-R110)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L130-R135)
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L130-R135):
Removed the `config-file` parameter from the `Initialize CodeQL` step,
as the associated configuration file is no longer required.