chore: Update Dependabot Schedule (#70)
# Pull Request

## Description

This pull request updates the Dependabot configuration to use a
cron-based schedule for dependency updates instead of the previous daily
interval. The changes ensure more precise control over when updates are
triggered.

**Dependabot schedule changes:**

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L10-R11):
Updated the `github-actions` dependency group to use a cron schedule
(`30 7 * * *`) instead of a daily interval.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L27-R27):
Updated the `typescript` dependency group to use a cron schedule (`30 7
* * *`) instead of a daily interval.