chore: Update Dependabot Groups and Configuration (#71)
# Pull Request

## Description

This pull request updates the `.github/dependabot.yml` configuration to
refine the behavior of dependency updates for GitHub Actions and
TypeScript. The changes include improvements to the scheduling,
grouping, and targeting of updates.

### Updates to GitHub Actions dependency configuration:
* Consolidated `directories` field into a single `directory` field with
the value `/`.
* Added a `timezone` field to the update schedule, specifying
`Europe/London` for more precise scheduling.
* Introduced the `applies-to` field in the `github-actions` group, with
the value `version-updates`, to better scope the updates.
* Added `exclude-patterns` to the `github-actions` group to prevent
updates for paths matching `super-linter/*`.

### Updates to TypeScript dependency configuration:
* Added the `timezone` field to the schedule, specifying
`Europe/London`.
* Introduced the `applies-to` field in the `typescript` group, with the
value `version-updates`, to align with the GitHub Actions configuration.