chore: Remove `.editorconfig-checker.json` (#98)
# Pull Request

## Description

This pull request includes two changes: the removal of the
`.editorconfig-checker.json` configuration file and an update to the
`lefthook.yml` file to specify a newer minimum version of Lefthook.

Configuration updates:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L10):
Removed the configuration file, likely because it is no longer needed or
used.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` from `1.11.11` to `1.11.13` to ensure
compatibility with a newer version of Lefthook.