deps(github-actions): Update Reusable Workflows (#119)
# Pull Request

## Description

This pull request updates several GitHub Actions workflows to use the
latest version of reusable workflows and modifies permissions for
improved functionality. Additionally, it removes an unused command from
the `Justfile`.

### Updates to GitHub Actions workflows:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow version for cache cleaning to
`v2025.06.06.01`.
* `.github/workflows/code-checks.yml`: 
- Updated the reusable workflow version for code checks and CodeQL
analysis to `v2025.06.06.01`.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L59-R62)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L73-R74)
- Adjusted permissions to include `actions: read` and `pull-requests:
write` for better compatibility.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL24-R24):
Updated the reusable workflow version for pull request tasks to
`v2025.06.06.01`.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18):
Updated the reusable workflow version for syncing labels to
`v2025.06.06.01`.

### Cleanup in `Justfile`:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL55-L58):
Removed the `zizmor-check-sarif` command, which generated SARIF output,
as it is no longer needed.